いったん敵1体とプレイヤーの居る状態は完成、敵の複数体化を実装中

+++ PortfolioGame/PortfolioGame/Utility/Vec3.h
@@ -19,7 +19,7 @@ public:
 	//ƒxƒNƒgƒ‹‚Ì’·‚³‚ÌŽæ“¾
 	inline float Length() const
 	{
-		float length = sqrtf(x * x + y * y + z * z);
+		float length = std::sqrt(x * x + y * y + z * z);
 		return length;
 	}
